Move defaultDirectives to src/helpers
